Use the latest version of Node.js (#72)
This fixes failures due to packages not supporting Node 14.x.

Additionally, fix the CI script for the `node-local-program-tsnode` to install dependencies for the program. This has apparently [never worked in CI](https://github.com/pulumi/automation-api-examples/actions/runs/7991360249/job/21822347128#step:10:40), but it wasn't previously reported as a failure. Now with the Node.js upgrade, it's failing in CI.

Finally, fix an actual problem with the `node-local-program-tsnode` program. It was failing with an "Access Denied" error trying to create the bucket policy. Fix this up to use the latest way of doing it.